requirements.txt
asgiref==3.3.1
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.13
django-ckeditor==6.0.0
django-embed-video==1.4.0
django-jet-reboot==1.1.1
django-js-asset==1.2.2
django-shapeshifter==18.9.23
gunicorn==20.1.0
idna==2.10
Pillow==8.3.0
pytz==2020.5
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
psycopg2==2.9.1
mysqlclient==2.0.3
